RatelimitChecker/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.1.0
packaging==26.1
pluggy==1.6.0
Pygments==2.20.0
pytest==8.4.2
pytest-mock==3.15.1
PyYAML==6.0.3
requests==2.32.5
tomli==2.4.1
typing_extensions==4.15.0
urllib3==2.6.3
